+++ src/refactoring/hooks/use-new-coupon.ts
@@ -1,0 +1,34 @@
+import { useState } from "react";
+
+import { CouponType } from "../../types";
+
+interface UseNewCouponProps {
+  onCouponAdd: (newCoupon: CouponType) => void;
+}
+
+const useNewCoupon = ({ onCouponAdd }: UseNewCouponProps) => {
+  const [newCoupon, setNewCoupon] = useState<CouponType>({
+    name: "",
+    code: "",
+    discountType: "percentage",
+    discountValue: 0,
+  });
+
+  const hanldeChangeCouponInfo = <K extends keyof CouponType>(key: K, value: CouponType[K]) => {
+    setNewCoupon(prev => ({ ...prev, [key]: value }));
+  };
+
+  const handleAddCoupon = () => {
+    onCouponAdd(newCoupon);
+    setNewCoupon({
+      name: "",
+      code: "",
+      discountType: "percentage",
+      discountValue: 0,
+    });
+  };
+
+  return { newCoupon, hanldeChangeCouponInfo, handleAddCoupon };
+};
+
+export default useNewCoupon;
